WEB-1179 fix Dockerfile for local dev
As it turns out the base docker image used (the "onbuild" variant) is
now deprecated and recommended to not be used.  See:

https://hub.docker.com/_/ruby/
https://github.com/docker-library/official-images/issues/2076

I found with the 2.3-onbuild image I'd get failure to even complete
the first line of the Dockerfile on a docker build.  This replaces
the onbuild image with a proper Ruby 2.3 image and modifies the
instructions for running as appropriate.